Refactor: move shadowban/event-ban enforcement into services
Address the thermo-nuclear review:

- Enforcement reads (is-banned / hidden-presenters / banned-users) now live
  on ShadowbanService + EventBanService; the views call request.services
  instead of hand-rolling Shadowban/EventBan ORM. Drops the raw model
  imports and the duplicated event-ban check from the views.
- Move ShadowbanRepository + EventBanRepository out of the 3k-line
  repositories.py into links/db/django/safety.py, and dedupe the
  username/email lookup into one _resolve_user helper.
- Build the event-ban "fake-full" view by constructing a fresh SessionData
  (dataclasses.replace) in adapters/web/django/safety_presentation.py
  instead of mutating the DTO in place in the view; simulacra use negative
  (clearly synthetic) pks.
- Avatar red ring is now a boolean `danger_ring` flag (single source of
  truth) instead of a copy-pasted Tailwind string in three templates.

Behavior-preserving; full suite green.
