+++ src/ludamus/pacts/safety.py
@@ -62,6 +62,10 @@ class ShadowbanRepositoryProtocol(Protocol):
     @staticmethod
     def list_candidates(owner_id: int) -> list[ShadowbanCandidateDTO]: ...
     @staticmethod
+    def banned_user_ids(owner_id: int) -> set[int]: ...
+    @staticmethod
+    def banning_owner_ids(target_id: int) -> set[int]: ...
+    @staticmethod
     def set_shadowban(*, owner_id: int, target_slug: str, banned: bool) -> None: ...
     @staticmethod
     def shadowban_by_identifier(*, owner_id: int, identifier: str) -> bool: ...
@@ -79,6 +83,8 @@ class EventBanRepositoryProtocol(Protocol):
     @staticmethod
     def list_by_event(event_id: int) -> list[EventBanDTO]: ...
     @staticmethod
+    def is_banned(*, event_id: int, user_id: int) -> bool: ...
+    @staticmethod
     def ban(*, event_id: int, identifier: str, reason: str) -> bool: ...
     @staticmethod
     def unban(*, event_id: int, ban_id: int) -> None: ...
@@ -86,6 +92,7 @@ class EventBanRepositoryProtocol(Protocol):
 
 class EventBanServiceProtocol(Protocol):
     def list_for_event(self, event_id: int) -> list[EventBanDTO]: ...
+    def is_banned(self, *, event_id: int, user_id: int) -> bool: ...
     def ban(self, *, event_id: int, identifier: str, reason: str) -> bool: ...
     def unban(self, *, event_id: int, ban_id: int) -> None: ...
 
@@ -98,6 +105,8 @@ class ShadowbanNotifierProtocol(Protocol):
 
 class ShadowbanServiceProtocol(Protocol):
     def list_candidates(self, owner_id: int) -> list[ShadowbanCandidateDTO]: ...
+    def banned_user_ids(self, owner_id: int) -> set[int]: ...
+    def banning_presenter_ids(self, viewer_id: int) -> set[int]: ...
     def set_shadowban(
         self, *, owner_id: int, target_slug: str, banned: bool
     ) -> None: ...
