+++ src/ludamus/templates/components/avatar.html
@@ -1,9 +1,9 @@
 {% load avatar_tags %}
-{# Reusable avatar: expects "user" (UserDTO/User) and optional "size", "ring_class" #}
+{# Reusable avatar: expects "user" (UserDTO/User) and optional "size", "danger_ring" #}
 {% with size_class=size|default:"size-8" border_class="border border-white/80 dark:border-neutral-900" %}
     {% if user %}
         {% with name_value=user.full_name|default:user.name|default:user.username|default:"" %}
-            <div class="{{ size_class }} {{ border_class }} {{ ring_class|default:'' }} rounded-full shrink-0 relative overflow-hidden">
+            <div class="{{ size_class }} {{ border_class }} {% if danger_ring|default:False %}ring-2 ring-danger{% endif %} rounded-full shrink-0 relative overflow-hidden">
                 <div class="absolute inset-0 {{ name_value|avatar_bg_class }} flex items-center justify-center">
                     <span class="text-xs font-semibold text-white">{{ name_value|slice:":2"|upper }}</span>
                 </div>
